feat: initial prototype scripts were pushed in the previous (1.1.1) release
This commit was made to document that the previous commit 54f51fb which is tagged as version 1.1.1 accidentally did not include an accurate commit message. The 1.1.1 release did include the deletion of testing files as mentioned in the commit message, but it also included the adding of all the scripts and instructions for the initial prototype, so I made this additional empty commit to document that.